Well, the name 'Arrays' seems to be not so good, since it collides with java.util.Arrays
